# frontend/src/api/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string._kind-string--path-string--value-any__": {
				"properties": {},
				"additionalProperties": {
					"properties": {
						"value": {},
						"path": {
							"type": "string"
						},
						"kind": {
							"type": "string"
						}
					},
					"required": [
						"value",
						"path",
						"kind"
					],
					"type": "object"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"DatabaseValidationFailResponse": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"$ref": "#/components/schemas/Record_string._kind-string--path-string--value-any__"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"DATABASE_VALIDATION_ERROR"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"FailResponse_NO_TOKEN_PROVIDED._reason-string__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"reason": {
										"type": "string"
									}
								},
								"required": [
									"reason"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"NO_TOKEN_PROVIDED"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_NoTokenProvidedError_": {
				"$ref": "#/components/schemas/FailResponse_NO_TOKEN_PROVIDED._reason-string__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"FailResponse_INVALID_TOKEN.undefined_": {
				"properties": {
					"data": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"INVALID_TOKEN"
								],
								"nullable": false
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_InvalidTokenError_": {
				"$ref": "#/components/schemas/FailResponse_INVALID_TOKEN.undefined_",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"FailResponse_TOKEN_USER_NOT_FOUND._userId-string__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"userId": {
										"type": "string"
									}
								},
								"required": [
									"userId"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"TOKEN_USER_NOT_FOUND"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_TokenUserNotFoundError_": {
				"$ref": "#/components/schemas/FailResponse_TOKEN_USER_NOT_FOUND._userId-string__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"FailResponse_AUTH_VERSION_MISMATCH._userId-string--currentVersion-number--tokenVersion-number__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"tokenVersion": {
										"type": "number",
										"format": "double"
									},
									"currentVersion": {
										"type": "number",
										"format": "double"
									},
									"userId": {
										"type": "string"
									}
								},
								"required": [
									"tokenVersion",
									"currentVersion",
									"userId"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"AUTH_VERSION_MISMATCH"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_AuthenticationVersionMismatchError_": {
				"$ref": "#/components/schemas/FailResponse_AUTH_VERSION_MISMATCH._userId-string--currentVersion-number--tokenVersion-number__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"AuthFailResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/FailResponseFromError_NoTokenProvidedError_"
					},
					{
						"$ref": "#/components/schemas/FailResponseFromError_InvalidTokenError_"
					},
					{
						"$ref": "#/components/schemas/FailResponseFromError_TokenUserNotFoundError_"
					},
					{
						"$ref": "#/components/schemas/FailResponseFromError_AuthenticationVersionMismatchError_"
					}
				],
				"description": "Tipo de respuesta para errores de autenticación JWT (401).\nCubre todos los errores lanzados por expressAuthentication:\n- NO_TOKEN_PROVIDED: No se proporcionó token (con reason en details)\n- INVALID_TOKEN: Token inválido o expirado\n- TOKEN_USER_NOT_FOUND: Usuario del token ya no existe (con userId en details)\n- AUTH_VERSION_MISMATCH: Sesión invalidada (con userId, currentVersion, tokenVersion en details)"
			},
...
			"SuccessResponse": {
				"description": "Wrapper de éxito JSend (aplicado por el middleware global).",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success"
						],
						"nullable": false
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponseData": {
				"description": "Respuesta del endpoint POST /auth/login",
				"properties": {
					"userId": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"authVersion": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"token",
					"authVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_LoginResponseData_": {
				"description": "Wrapper de éxito JSend (aplicado por el middleware global).",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success"
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/LoginResponseData"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationDetails_body-or-body.identifier-or-body.password-or-body.responseType_": {
				"properties": {
					"body": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"body.password": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"body.identifier": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"body.responseType": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				},
				"type": "object",
				"description": "Generador de estructura de detalles.\nTransforma una unión de llaves ('body' | 'body.id') en el objeto que TSOA espera."
			},
			"RequestValidationFailResponse_ValidationDetails_body-or-body.identifier-or-body.password-or-body.responseType__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"$ref": "#/components/schemas/ValidationDetails_body-or-body.identifier-or-body.password-or-body.responseType_"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"REQUEST_VALIDATION_ERROR"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"LoginRequestValidationFailResponse": {
				"$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_body-or-body.identifier-or-body.password-or-body.responseType__"
			},
			"LoginValidationFailResponse": {
				"$ref": "#/components/schemas/LoginRequestValidationFailResponse"
			},
			"FailResponse_INVALID_CREDENTIALS._identifier-string__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"identifier": {
										"type": "string"
									}
								},
								"required": [
									"identifier"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"INVALID_CREDENTIALS"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_InvalidCredentialsError_": {
				"$ref": "#/components/schemas/FailResponse_INVALID_CREDENTIALS._identifier-string__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"LoginRequest": {
				"properties": {
					"identifier": {
						"type": "string",
						"minLength": 3
					},
					"password": {
						"type": "string",
						"minLength": 8
					},
					"responseType": {
						"type": "string",
						"enum": [
							"cookie",
							"json"
						],
						"description": "Indicates how the token should be returned.",
						"default": "json"
					}
				},
				"required": [
					"identifier",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageResponseData": {
				"description": "Respuesta de éxito solo con mensaje (sin datos).\nEl wrapper global lo convertirá en: { status: 'success', data: { message: string } }",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_MessageResponseData_": {
				"description": "Wrapper de éxito JSend (aplicado por el middleware global).",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success"
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/MessageResponseData"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationDetails_body-or-body.oldPassword-or-body.newPassword_": {
				"properties": {
					"body": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"body.oldPassword": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"body.newPassword": {
						"properties": {
							"value": {},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				},
				"type": "object",
				"description": "Generador de estructura de detalles.\nTransforma una unión de llaves ('body' | 'body.id') en el objeto que TSOA espera."
			},
			"RequestValidationFailResponse_ValidationDetails_body-or-body.oldPassword-or-body.newPassword__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"$ref": "#/components/schemas/ValidationDetails_body-or-body.oldPassword-or-body.newPassword_"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"REQUEST_VALIDATION_ERROR"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ChangePasswordRequestValidationFailResponse": {
				"$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_body-or-body.oldPassword-or-body.newPassword__"
			},
			"ChangePasswordValidationFailResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ChangePasswordRequestValidationFailResponse"
					},
					{
						"$ref": "#/components/schemas/DatabaseValidationFailResponse"
					}
				]
			},
			"FailResponse_INVALID_PASSWORD._identifier-string__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"identifier": {
										"type": "string"
									}
								},
								"required": [
									"identifier"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"INVALID_PASSWORD"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_InvalidPasswordError_": {
				"$ref": "#/components/schemas/FailResponse_INVALID_PASSWORD._identifier-string__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"FailResponse_USER_NOT_FOUND._identifier-string__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"properties": {
									"identifier": {
										"type": "string"
									}
								},
								"required": [
									"identifier"
								],
								"type": "object"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"USER_NOT_FOUND"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_LoginUserNotFoundError_": {
				"$ref": "#/components/schemas/FailResponse_USER_NOT_FOUND._identifier-string__",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"ChangePasswordRequest": {
				"properties": {
					"oldPassword": {
						"type": "string",
						"minLength": 8
					},
					"newPassword": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": [
					"oldPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RequestValidationFailResponse_ValidationDetails_body-or-body.email__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"$ref": "#/components/schemas/ValidationDetails_body-or-body.email_"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"REQUEST_VALIDATION_ERROR"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ForgotPasswordRequestValidationFailResponse": {
				"$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_body-or-body.email__"
			},
			"ForgotPasswordValidationFailResponse": {
				"$ref": "#/components/schemas/ForgotPasswordRequestValidationFailResponse"
			},
			"ForgotPasswordRequest": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RequestValidationFailResponse_ValidationDetails_body-or-body.token-or-body.newPassword__": {
				"properties": {
					"data": {
						"properties": {
							"details": {
								"$ref": "#/components/schemas/ValidationDetails_body-or-body.token-or-body.newPassword_"
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"REQUEST_VALIDATION_ERROR"
								],
								"nullable": false
							}
						},
						"required": [
							"details",
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ResetPasswordRequestValidationFailResponse": {
				"$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_body-or-body.token-or-body.newPassword__"
			},
			"ResetPasswordValidationFailResponse": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ResetPasswordRequestValidationFailResponse"
					},
					{
						"$ref": "#/components/schemas/DatabaseValidationFailResponse"
					}
				]
			},
			"FailResponse_RESET_TOKEN_INVALID_OR_EXPIRED.undefined_": {
				"properties": {
					"data": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string",
								"enum": [
									"RESET_TOKEN_INVALID_OR_EXPIRED"
								],
								"nullable": false
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"fail"
						],
						"nullable": false
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object",
				"description": "Tipo genérico base para respuestas de error con código y detalles opcionales."
			},
			"FailResponseFromError_ResetTokenInvalidOrExpiredError_": {
				"$ref": "#/components/schemas/FailResponse_RESET_TOKEN_INVALID_OR_EXPIRED.undefined_",
				"description": "Extrae automáticamente code y details de una clase AppError.\n\nEjemplo:\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
			},
			"ResetPasswordRequest": {
				"properties": {
					"token": {
						"type": "string"
					},
					"newPassword": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": [
					"token",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"jwt-optional": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@tfg/backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse_LoginResponseData_"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Credenciales inválidas",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponseFromError_InvalidCredentialsError_"
								}
							}
						}
					},
					"422": {
						"description": "Error de validación",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginValidationFailResponse"
								}
							}
						}
					}
				},
				"description": "Inicia sesión con un identificador (email o username) y contraseña.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
...
		"/auth/logoutAll": {
			"post": {
				"operationId": "LogoutAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "No autenticado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthFailResponse"
								}
							}
						}
					}
				},
				"description": "Cierra todas las sesiones activas del usuario (invalida todos los tokens).",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
								}
							}
						}
					},
					"401": {
						"description": "Contraseña incorrecta",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponseFromError_InvalidPasswordError_"
								}
							}
						}
					},
					"404": {
						"description": "Usuario no encontrado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponseFromError_LoginUserNotFoundError_"
								}
							}
						}
					},
					"422": {
						"description": "Error de validación",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChangePasswordValidationFailResponse"
								}
							}
						}
					}
				},
				"description": "Cambia la contraseña del usuario autenticado.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							}
						}
					}
				}
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
								}
							}
						}
					},
					"422": {
						"description": "Error de validación",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForgotPasswordValidationFailResponse"
								}
							}
						}
					}
				},
				"description": "Solicita un email de recuperación de contraseña.\nPor seguridad, siempre devuelve éxito aunque el email no exista.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
								}
							}
						}
					},
					"400": {
						"description": "Token inválido o expirado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponseFromError_ResetTokenInvalidOrExpiredError_"
								}
							}
						}
					},
					"422": {
						"description": "Error de validación",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordValidationFailResponse"
								}
							}
						}
					}
				},
				"description": "Restablece la contraseña usando un token de recuperación.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}